requirements.txt
uvicorn==0.27.0
fastapi==0.109.0
firebase-admin

# Core Google Gemini client
google-genai==1.9.0
google-generativeai

# Data handling
pandas
openpyxl

# Embeddings & ML
sentence-transformers
torch

# Pinecone vector database
pinecone

# BigQuery client
google-cloud-bigquery

# Geolocation lookup
# googlemaps==4.10.0

# Date parsing
dateparser

# Spacy
#spacy>=3.0.0
#en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl

# US Address
# usaddress